backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-genai
python-dotenv
dj-database-url
psycopg2-binary
whitenoise
gunicorn
